+++ Net_Core/DataAccess/Interfaces/ILogging.cs
@@ -2,6 +2,7 @@ using GrowthWare.Framework.Models;
 using GrowthWare.DataAccess.Interfaces.Base;
 using System;
 using System.Data;
+using System.Threading.Tasks;
 
 namespace GrowthWare.DataAccess.Interfaces
 {
@@ -16,12 +17,12 @@ namespace GrowthWare.DataAccess.Interfaces
         /// </summary>
         /// <param name="logSeqId">int</param>
         /// <returns>MLoggingProfile</returns>
-        MLoggingProfile GetLog(int logSeqId);
+        Task<MLoggingProfile> GetLog(int logSeqId);
 
         /// <summary>
         /// Saves this instance.
         /// </summary>
-        void Save(MLoggingProfile profile);
+        Task Save(MLoggingProfile profile);
     }
 
 }