+++ Net_Core/DataAccess/SQLServer/DLogging.cs
@@ -4,6 +4,7 @@ using GrowthWare.Framework.Models;
 using System;
 using System.Data;
 using System.Data.SqlClient;
+using System.Threading.Tasks;
 
 namespace GrowthWare.DataAccess.SQLServer;
 public class DLogging : AbstractDBInteraction, ILogging
@@ -16,14 +17,14 @@ public class DLogging : AbstractDBInteraction, ILogging
     }
 #endregion
 
-    MLoggingProfile ILogging.GetLog(int logSeqId)
+    async Task<MLoggingProfile> ILogging.GetLog(int logSeqId)
     {
         String mStoredProcedure = "[ZGWSystem].[Get_Logs]";
         SqlParameter[] mParameters =
         {
            new SqlParameter("@P_LogSeqId", logSeqId)
         };
-        DataRow mDataRow = base.GetDataRow(mStoredProcedure, mParameters);
+        DataRow mDataRow = await base.GetDataRowAsync(mStoredProcedure, mParameters);
         if(mDataRow != null)
         {
             return new MLoggingProfile(mDataRow);
@@ -34,7 +35,7 @@ public class DLogging : AbstractDBInteraction, ILogging
         }
     }
 
-    void ILogging.Save(MLoggingProfile profile)
+    async Task ILogging.Save(MLoggingProfile profile)
     {
         if (profile == null) throw new ArgumentNullException(nameof(profile), "profile cannot be a null reference (Nothing in Visual Basic)!");
         String mStoredProcedure = "[ZGWSystem].[Set_Log]";
@@ -48,6 +49,6 @@ public class DLogging : AbstractDBInteraction, ILogging
             new SqlParameter("@P_Msg", profile.Msg),
             base.GetSqlParameter("@P_Primary_Key", 0, ParameterDirection.Output)
         };
-        base.ExecuteNonQuery(mStoredProcedure, mParameters);
+        await base.ExecuteNonQueryAsync(mStoredProcedure, mParameters);
     }
 }